requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-debug-toolbar==3.5.0
drf-nested-routers==0.93.4
django-filter==22.1
djoser==2.1.0
djangorestframework_simplejwt==4.8.0